Backend/requirements.txt
alembic
bcrypt
fastapi
httpx
passlib[bcrypt]
pydantic
pydantic-settings
pytest
pytest-cov
pytest-mock
python-dotenv
python-jose[cryptography]
python-multipart
psycopg2-binary
requests
sqlalchemy
stripe
uvicorn[standart]
validate_docbr